Commands now in object
I put the interface() commands into an object as an index so we won't need to type 'if input ==...' all the time. I used 'try:' to determine if the input is a valid command. If it is, it responds. If not, it does nothing; we don't want the bot constantly interrupting everybody with 'command not found' every time they say something.